Revert "Refactor loading and determining unread entries"
This reverts commit 4a7f588bdbe5181d05818449ddc711e53390d5fc.

I'm not really happy about this, however there's a couple reasons for
it. Anytime you're doing the query with a LIMIT and an ORDER BY postgres
(and probably mysql) does not have a good time. Additionally, I was
doing this to cut down from two queries to one, and not remembering that
the total unread count was used elsewhere in the views.

So rolling this back will create two queries for showing the reading
list instead of one - first one query to get all unread entry IDs, and
then a second query to actually fetch the content for a limited amount
of entries.